From 83bc93a1cfa5f492a7238f284a61d000971a44db Mon Sep 17 00:00:00 2001 Subject: [PATCH 215/418] sched: Add saved_state for tasks blocked on sleeping  locks
Spinlocks are state preserving in !RT. RT changes the state when a
task gets blocked on a lock. So we need to remember the state before
the lock contention. If a regular wakeup (not a RTmutex related
wakeup) happens, the saved_state is updated to running. When the lock
sleep is done, the saved state is restored.

Signed-off-by: Thomas Gleixner <tglx@linutronix.de>