From 0bc7b41c9d6adcc1aa2a03d7b705557e85155153 Mon Sep 17 00:00:00 2001 Subject: [PATCH 223/418] sched: ttwu: Return success when only changing the  saved_state value
When a task blocks on a rt lock, it saves the current state in
p->saved_state, so a lock related wake up will not destroy the
original state.

When a real wakeup happens, while the task is running due to a lock
wakeup already, we update p->saved_state to TASK_RUNNING, but we do
not return success, which might cause another wakeup in the waitqueue
code and the task remains in the waitqueue list. Return success in
that case as well.

Signed-off-by: Thomas Gleixner <tglx@linutronix.de>