From fa87e7c559db3670cfb5ac4716b84cd2cf658e9d Mon Sep 17 00:00:00 2001 Subject: [PATCH 352/418] mm: rt: Fix generic kmap_atomic for RT
The update to 4.1 brought in the mainline variant of the pagefault
disable distangling from preempt count. That introduced a
preempt_disable/enable pair in the generic kmap_atomic/kunmap_atomic
implementations which got not converted to the _nort() variant.

That results in massive 'scheduling while atomic/sleeping function
called from invalid context' splats.

Fix that up.

Reported-and-tested-by: Juergen Borleis <jbe@pengutronix.de>
Signed-off-by: Thomas Gleixner <tglx@linutronix.de>