From 14ff530effdb2c48c594ef3507999cc5b9618e48 Mon Sep 17 00:00:00 2001 Subject: [PATCH 336/418] kernel/printk: Don't try to print from IRQ/NMI region
On -RT we try to acquire sleeping locks which might lead to warnings
from lockdep or a warn_on() from spin_try_lock() (which is a rtmutex on
RT).
We don't print in general from a IRQ off region so we should not try
this via console_unblank() / bust_spinlocks() as well.

Signed-off-by: Sebastian Andrzej Siewior <bigeasy@linutronix.de>