From 2ab3b1d3828aed39e97cdb4a1cdd6d368873c1a9 Mon Sep 17 00:00:00 2001 Subject: [PATCH 242/418] fs/nfs: turn rmdir_sem into a semaphore
The RW semaphore had a reader side which used the _non_owner version
because it most likely took the reader lock in one thread and released it
in another which would cause lockdep to complain if the "regular"
version was used.
On -RT we need the owner because the rw lock is turned into a rtmutex.
The semaphores on the hand are "plain simple" and should work as
expected. We can't have multiple readers but on -RT we don't allow
multiple readers anyway so that is not a loss.

Signed-off-by: Sebastian Andrzej Siewior <bigeasy@linutronix.de>