Subject: [PATCH] RCU: we need to skip that warning but only on sleeping  locks
This check is okay for upstream. On RT we trigger this while blocking on
sleeping lock. In this case, it is okay to schedule() within a RCU
section.
Since spin_lock() and read_lock() disables migration it should be okay
to test for this as an indication whether or not a sleeping lock is
held. The ->pi_blocked_on member won't work becasuse it might also be
set on regular mutexes.

Signed-off-by: Sebastian Andrzej Siewior <bigeasy@linutronix.de>