From 64f7609a4afaa3ae86b180e6b390ab61c7297756 Mon Sep 17 00:00:00 2001 Subject: [PATCH 065/418] drm/i915: properly init lockdep class
The code has an ifdef and uses two functions to either init the bare
spinlock or init it and set a lock-class. It is possible to do the same
thing without an ifdef.
With this patch (in debug case) we first use the "default" lock class
which is later overwritten to the supplied one. Without lockdep the set
name/class function vanishes.

Signed-off-by: Sebastian Andrzej Siewior <bigeasy@linutronix.de>