From ddebd4ab0ba2bc665aa7648dd78e99b277980d2c Mon Sep 17 00:00:00 2001 Subject: [PATCH 113/418] tracing: Add a clock attribute for hist triggers
The default clock if timestamps are used in a histogram is "global".
If timestamps aren't used, the clock is irrelevant.

Use the "clock=" param only if you want to override the default
"global" clock for a histogram with timestamps.

Link: http://lkml.kernel.org/r/427bed1389c5d22aa40c3e0683e30cc3d151e260.1516069914.git.tom.zanussi@linux.intel.com

Signed-off-by: Tom Zanussi <tom.zanussi@linux.intel.com>
Signed-off-by: Rajvi Jingar <rajvi.jingar@intel.com>
Signed-off-by: Steven Rostedt (VMware) <rostedt@goodmis.org>
(cherry picked from commit 77e7689e0b182465cfcd7c328061b70eecdcde31)
Signed-off-by: Sebastian Andrzej Siewior <bigeasy@linutronix.de>