From a4b0759b36f87efae28313c89ce17f46abb7e203 Mon Sep 17 00:00:00 2001 Subject: [PATCH 116/418] ring-buffer: Add nesting for adding events within  events
The ring-buffer code has recusion protection in case tracing ends up tracing
itself, the ring-buffer will detect that it was called at the same context
(normal, softirq, interrupt or NMI), and not continue to record the event.

With the histogram synthetic events, they are called while tracing another
event at the same context. The recusion protection triggers because it
detects tracing at the same context and stops it.

Add ring_buffer_nest_start() and ring_buffer_nest_end() that will notify the
ring buffer that a trace is about to happen within another trace and that it
is intended, and not to trigger the recursion blocking.

Signed-off-by: Steven Rostedt (VMware) <rostedt@goodmis.org>
(cherry picked from commit f932ff1d98c482716b4b71a5d76b2aa3d65f66f0)
Signed-off-by: Sebastian Andrzej Siewior <bigeasy@linutronix.de>